fixed ssr error with marcas routes
