+++ .vercel/output/_functions/manifest_DNVqlSOt.mjs
@@ -1,0 +1,102 @@
+import 'cookie';
+import 'kleur/colors';
+import { N as NOOP_MIDDLEWARE_FN } from './chunks/astro-designed-error-pages_C6FDl1jh.mjs';
+import 'es-module-lexer';
+import { f as decodeKey } from './chunks/astro/server_CSx5QtHz.mjs';
+import 'clsx';
+
+function sanitizeParams(params) {
+  return Object.fromEntries(
+    Object.entries(params).map(([key, value]) => {
+      if (typeof value === "string") {
+        return [key, value.normalize().replace(/#/g, "%23").replace(/\?/g, "%3F")];
+      }
+      return [key, value];
+    })
+  );
+}
+function getParameter(part, params) {
+  if (part.spread) {
+    return params[part.content.slice(3)] || "";
+  }
+  if (part.dynamic) {
+    if (!params[part.content]) {
+      throw new TypeError(`Missing parameter: ${part.content}`);
+    }
+    return params[part.content];
+  }
+  return part.content.normalize().replace(/\?/g, "%3F").replace(/#/g, "%23").replace(/%5B/g, "[").replace(/%5D/g, "]");
+}
+function getSegment(segment, params) {
+  const segmentPath = segment.map((part) => getParameter(part, params)).join("");
+  return segmentPath ? "/" + segmentPath : "";
+}
+function getRouteGenerator(segments, addTrailingSlash) {
+  return (params) => {
+    const sanitizedParams = sanitizeParams(params);
+    let trailing = "";
+    if (addTrailingSlash === "always" && segments.length) {
+      trailing = "/";
+    }
+    const path = segments.map((segment) => getSegment(segment, sanitizedParams)).join("") + trailing;
+    return path || "/";
+  };
+}
+
+function deserializeRouteData(rawRouteData) {
+  return {
+    route: rawRouteData.route,
+    type: rawRouteData.type,
+    pattern: new RegExp(rawRouteData.pattern),
+    params: rawRouteData.params,
+    component: rawRouteData.component,
+    generate: getRouteGenerator(rawRouteData.segments, rawRouteData._meta.trailingSlash),
+    pathname: rawRouteData.pathname || void 0,
+    segments: rawRouteData.segments,
+    prerender: rawRouteData.prerender,
+    redirect: rawRouteData.redirect,
+    redirectRoute: rawRouteData.redirectRoute ? deserializeRouteData(rawRouteData.redirectRoute) : void 0,
+    fallbackRoutes: rawRouteData.fallbackRoutes.map((fallback) => {
+      return deserializeRouteData(fallback);
+    }),
+    isIndex: rawRouteData.isIndex,
+    origin: rawRouteData.origin
+  };
+}
+
+function deserializeManifest(serializedManifest) {
+  const routes = [];
+  for (const serializedRoute of serializedManifest.routes) {
+    routes.push({
+      ...serializedRoute,
+      routeData: deserializeRouteData(serializedRoute.routeData)
+    });
+    const route = serializedRoute;
+    route.routeData = deserializeRouteData(serializedRoute.routeData);
+  }
+  const assets = new Set(serializedManifest.assets);
+  const componentMetadata = new Map(serializedManifest.componentMetadata);
+  const inlinedScripts = new Map(serializedManifest.inlinedScripts);
+  const clientDirectives = new Map(serializedManifest.clientDirectives);
+  const serverIslandNameMap = new Map(serializedManifest.serverIslandNameMap);
+  const key = decodeKey(serializedManifest.key);
+  return {
+    // in case user middleware exists, this no-op middleware will be reassigned (see plugin-ssr.ts)
+    middleware() {
+      return { onRequest: NOOP_MIDDLEWARE_FN };
+    },
+    ...serializedManifest,
+    assets,
+    componentMetadata,
+    inlinedScripts,
+    clientDirectives,
+    routes,
+    serverIslandNameMap,
+    key
+  };
+}
+
+const manifest = deserializeManifest({"hrefRoot":"file:///home/dan/all-for-water/","adapterName":"@astrojs/vercel","routes":[{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"type":"page","component":"_server-islands.astro","params":["name"],"segments":[[{"content":"_server-islands","dynamic":false,"spread":false}],[{"content":"name","dynamic":true,"spread":false}]],"pattern":"^\\/_server-islands\\/([^/]+?)\\/?$","prerender":false,"isIndex":false,"fallbackRoutes":[],"route":"/_server-islands/[name]","origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"inline","content":"[data-astro-image]{width:100%;height:auto;-o-object-fit:var(--fit);object-fit:var(--fit);-o-object-position:var(--pos);object-position:var(--pos);aspect-ratio:var(--w) / var(--h)}[data-astro-image=responsive]{max-width:calc(var(--w) * 1px);max-height:calc(var(--h) * 1px)}[data-astro-image=fixed]{width:calc(var(--w) * 1px);height:calc(var(--h) * 1px)}\n"}],"routeData":{"type":"endpoint","isIndex":false,"route":"/_image","pattern":"^\\/_image\\/?$","segments":[[{"content":"_image","dynamic":false,"spread":false}]],"params":[],"component":"node_modules/astro/dist/assets/endpoint/generic.js","pathname":"/_image","prerender":false,"fallbackRoutes":[],"origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/_id_.Dhwt1lP-.css"},{"type":"inline","content":"@import\"https://fonts.googleapis.com/css2?family=Edu+AU+VIC+WA+NT+Pre:wght@400..700&display=swap\";.nav-links[data-astro-cid-3ef6ksr2]{font-size:1.2rem;font-weight:500;width:-moz-fit-content;width:fit-content}.nav-links[data-astro-cid-3ef6ksr2]:after{content:\"\";display:block;width:0;height:2.5px;background-color:#fff;transition:all .25s cubic-bezier(.645,.045,.355,1) 0s;opacity:1;z-index:101}.nav-links[data-astro-cid-3ef6ksr2]:hover:after{width:100%}.fade-enter[data-astro-cid-3ef6ksr2]{opacity:0;transform:translateY(-10px);animation:fadeIn .3s forwards}.fade-exit[data-astro-cid-3ef6ksr2]{opacity:1;transform:translateY(0);animation:fadeOut .3s forwards}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10px)}to{opacity:1;transform:translateY(0)}}@keyframes fadeOut{0%{opacity:1;transform:translateY(0)}to{opacity:0;transform:translateY(-10px)}}\n.astro-route-announcer{position:absolute;left:0;top:0;clip:rect(0 0 0 0);clip-path:inset(50%);overflow:hidden;white-space:nowrap;width:1px;height:1px}\n"}],"routeData":{"route":"/contact","isIndex":false,"type":"page","pattern":"^\\/contact\\/?$","segments":[[{"content":"contact","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/contact.astro","pathname":"/contact","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/_id_.Dhwt1lP-.css"},{"type":"inline","content":".astro-route-announcer{position:absolute;left:0;top:0;clip:rect(0 0 0 0);clip-path:inset(50%);overflow:hidden;white-space:nowrap;width:1px;height:1px}\n@import\"https://fonts.googleapis.com/css2?family=Edu+AU+VIC+WA+NT+Pre:wght@400..700&display=swap\";.nav-links[data-astro-cid-rvkblzjp]{font-size:1.2rem;font-weight:500;width:-moz-fit-content;width:fit-content}.nav-links[data-astro-cid-rvkblzjp]:after{content:\"\";display:block;width:0;height:2.5px;background-color:#fff;transition:all .25s cubic-bezier(.645,.045,.355,1) 0s;opacity:1;z-index:101}.nav-links[data-astro-cid-rvkblzjp]:hover:after{width:100%}.fade-enter[data-astro-cid-rvkblzjp]{opacity:0;transform:translateY(-10px);animation:fadeIn .3s forwards}.fade-exit[data-astro-cid-rvkblzjp]{opacity:1;transform:translateY(0);animation:fadeOut .3s forwards}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10px)}to{opacity:1;transform:translateY(0)}}@keyframes fadeOut{0%{opacity:1;transform:translateY(0)}to{opacity:0;transform:translateY(-10px)}}\n"}],"routeData":{"route":"/es/contacto","isIndex":false,"type":"page","pattern":"^\\/es\\/contacto\\/?$","segments":[[{"content":"es","dynamic":false,"spread":false}],[{"content":"contacto","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/es/contacto.astro","pathname":"/es/contacto","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/_id_.Dhwt1lP-.css"},{"type":"inline","content":"@import\"https://fonts.googleapis.com/css2?family=Edu+AU+VIC+WA+NT+Pre:wght@400..700&display=swap\";.nav-links[data-astro-cid-3ef6ksr2]{font-size:1.2rem;font-weight:500;width:-moz-fit-content;width:fit-content}.nav-links[data-astro-cid-3ef6ksr2]:after{content:\"\";display:block;width:0;height:2.5px;background-color:#fff;transition:all .25s cubic-bezier(.645,.045,.355,1) 0s;opacity:1;z-index:101}.nav-links[data-astro-cid-3ef6ksr2]:hover:after{width:100%}.fade-enter[data-astro-cid-3ef6ksr2]{opacity:0;transform:translateY(-10px);animation:fadeIn .3s forwards}.fade-exit[data-astro-cid-3ef6ksr2]{opacity:1;transform:translateY(0);animation:fadeOut .3s forwards}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10px)}to{opacity:1;transform:translateY(0)}}@keyframes fadeOut{0%{opacity:1;transform:translateY(0)}to{opacity:0;transform:translateY(-10px)}}\n.astro-route-announcer{position:absolute;left:0;top:0;clip:rect(0 0 0 0);clip-path:inset(50%);overflow:hidden;white-space:nowrap;width:1px;height:1px}\n"}],"routeData":{"route":"/es/marca/[marca]","isIndex":false,"type":"page","pattern":"^\\/es\\/marca\\/([^/]+?)\\/?$","segments":[[{"content":"es","dynamic":false,"spread":false}],[{"content":"marca","dynamic":false,"spread":false}],[{"content":"marca","dynamic":true,"spread":false}]],"params":["marca"],"component":"src/pages/es/marca/[marca].astro","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/_id_.Dhwt1lP-.css"},{"type":"inline","content":".astro-route-announcer{position:absolute;left:0;top:0;clip:rect(0 0 0 0);clip-path:inset(50%);overflow:hidden;white-space:nowrap;width:1px;height:1px}\n@import\"https://fonts.googleapis.com/css2?family=Edu+AU+VIC+WA+NT+Pre:wght@400..700&display=swap\";.nav-links[data-astro-cid-rvkblzjp]{font-size:1.2rem;font-weight:500;width:-moz-fit-content;width:fit-content}.nav-links[data-astro-cid-rvkblzjp]:after{content:\"\";display:block;width:0;height:2.5px;background-color:#fff;transition:all .25s cubic-bezier(.645,.045,.355,1) 0s;opacity:1;z-index:101}.nav-links[data-astro-cid-rvkblzjp]:hover:after{width:100%}.fade-enter[data-astro-cid-rvkblzjp]{opacity:0;transform:translateY(-10px);animation:fadeIn .3s forwards}.fade-exit[data-astro-cid-rvkblzjp]{opacity:1;transform:translateY(0);animation:fadeOut .3s forwards}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10px)}to{opacity:1;transform:translateY(0)}}@keyframes fadeOut{0%{opacity:1;transform:translateY(0)}to{opacity:0;transform:translateY(-10px)}}\nhtml{scroll-behavior:smooth}\n"}],"routeData":{"route":"/es/[id_es]","isIndex":false,"type":"page","pattern":"^\\/es\\/([^/]+?)\\/?$","segments":[[{"content":"es","dynamic":false,"spread":false}],[{"content":"id_es","dynamic":true,"spread":false}]],"params":["id_es"],"component":"src/pages/es/[id_es].astro","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/_id_.Dhwt1lP-.css"},{"type":"inline","content":"@keyframes slide-in-right{0%{transform:translate(100%);opacity:0}to{transform:translate(0);opacity:1}}@keyframes slide-out-left{0%{transform:translate(0);opacity:1}to{transform:translate(-100%);opacity:0}}.animate-slide-in-right[data-astro-cid-wgk6sxh5]{animation:slide-in-right .3s forwards}.animate-slide-out-left[data-astro-cid-wgk6sxh5]{animation:slide-out-left .3s forwards}\n.astro-route-announcer{position:absolute;left:0;top:0;clip:rect(0 0 0 0);clip-path:inset(50%);overflow:hidden;white-space:nowrap;width:1px;height:1px}\n@import\"https://fonts.googleapis.com/css2?family=Edu+AU+VIC+WA+NT+Pre:wght@400..700&display=swap\";.nav-links[data-astro-cid-rvkblzjp]{font-size:1.2rem;font-weight:500;width:-moz-fit-content;width:fit-content}.nav-links[data-astro-cid-rvkblzjp]:after{content:\"\";display:block;width:0;height:2.5px;background-color:#fff;transition:all .25s cubic-bezier(.645,.045,.355,1) 0s;opacity:1;z-index:101}.nav-links[data-astro-cid-rvkblzjp]:hover:after{width:100%}.fade-enter[data-astro-cid-rvkblzjp]{opacity:0;transform:translateY(-10px);animation:fadeIn .3s forwards}.fade-exit[data-astro-cid-rvkblzjp]{opacity:1;transform:translateY(0);animation:fadeOut .3s forwards}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10px)}to{opacity:1;transform:translateY(0)}}@keyframes fadeOut{0%{opacity:1;transform:translateY(0)}to{opacity:0;transform:translateY(-10px)}}\n"}],"routeData":{"route":"/es","isIndex":true,"type":"page","pattern":"^\\/es\\/?$","segments":[[{"content":"es","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/es/index.astro","pathname":"/es","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/_id_.Dhwt1lP-.css"},{"type":"inline","content":"@import\"https://fonts.googleapis.com/css2?family=Edu+AU+VIC+WA+NT+Pre:wght@400..700&display=swap\";.nav-links[data-astro-cid-3ef6ksr2]{font-size:1.2rem;font-weight:500;width:-moz-fit-content;width:fit-content}.nav-links[data-astro-cid-3ef6ksr2]:after{content:\"\";display:block;width:0;height:2.5px;background-color:#fff;transition:all .25s cubic-bezier(.645,.045,.355,1) 0s;opacity:1;z-index:101}.nav-links[data-astro-cid-3ef6ksr2]:hover:after{width:100%}.fade-enter[data-astro-cid-3ef6ksr2]{opacity:0;transform:translateY(-10px);animation:fadeIn .3s forwards}.fade-exit[data-astro-cid-3ef6ksr2]{opacity:1;transform:translateY(0);animation:fadeOut .3s forwards}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10px)}to{opacity:1;transform:translateY(0)}}@keyframes fadeOut{0%{opacity:1;transform:translateY(0)}to{opacity:0;transform:translateY(-10px)}}\n.astro-route-announcer{position:absolute;left:0;top:0;clip:rect(0 0 0 0);clip-path:inset(50%);overflow:hidden;white-space:nowrap;width:1px;height:1px}\nhtml{scroll-behavior:smooth}\n"}],"routeData":{"route":"/[id]","isIndex":false,"type":"page","pattern":"^\\/([^/]+?)\\/?$","segments":[[{"content":"id","dynamic":true,"spread":false}]],"params":["id"],"component":"src/pages/[id].astro","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"type":"external","value":"/_astro/page.BNYwb576.js"},{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/_id_.Dhwt1lP-.css"},{"type":"inline","content":"@keyframes slide-in-right{0%{transform:translate(100%);opacity:0}to{transform:translate(0);opacity:1}}@keyframes slide-out-left{0%{transform:translate(0);opacity:1}to{transform:translate(-100%);opacity:0}}.animate-slide-in-right[data-astro-cid-dnnxzc3h]{animation:slide-in-right .3s forwards}.animate-slide-out-left[data-astro-cid-dnnxzc3h]{animation:slide-out-left .3s forwards}\n@import\"https://fonts.googleapis.com/css2?family=Edu+AU+VIC+WA+NT+Pre:wght@400..700&display=swap\";.nav-links[data-astro-cid-3ef6ksr2]{font-size:1.2rem;font-weight:500;width:-moz-fit-content;width:fit-content}.nav-links[data-astro-cid-3ef6ksr2]:after{content:\"\";display:block;width:0;height:2.5px;background-color:#fff;transition:all .25s cubic-bezier(.645,.045,.355,1) 0s;opacity:1;z-index:101}.nav-links[data-astro-cid-3ef6ksr2]:hover:after{width:100%}.fade-enter[data-astro-cid-3ef6ksr2]{opacity:0;transform:translateY(-10px);animation:fadeIn .3s forwards}.fade-exit[data-astro-cid-3ef6ksr2]{opacity:1;transform:translateY(0);animation:fadeOut .3s forwards}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10px)}to{opacity:1;transform:translateY(0)}}@keyframes fadeOut{0%{opacity:1;transform:translateY(0)}to{opacity:0;transform:translateY(-10px)}}\n.astro-route-announcer{position:absolute;left:0;top:0;clip:rect(0 0 0 0);clip-path:inset(50%);overflow:hidden;white-space:nowrap;width:1px;height:1px}\n"}],"routeData":{"route":"/","isIndex":true,"type":"page","pattern":"^\\/$","segments":[],"params":[],"component":"src/pages/index.astro","pathname":"/","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}}],"site":"https://allforwater.com","base":"/","trailingSlash":"ignore","compressHTML":true,"componentMetadata":[["/home/dan/all-for-water/src/pages/es/[id_es].astro",{"propagation":"none","containsHead":true}],["/home/dan/all-for-water/src/pages/es/contacto.astro",{"propagation":"none","containsHead":true}],["/home/dan/all-for-water/src/pages/es/index.astro",{"propagation":"none","containsHead":true}],["/home/dan/all-for-water/src/pages/[id].astro",{"propagation":"none","containsHead":true}],["/home/dan/all-for-water/src/pages/contact.astro",{"propagation":"none","containsHead":true}],["/home/dan/all-for-water/src/pages/es/marca/[marca].astro",{"propagation":"none","containsHead":true}],["/home/dan/all-for-water/src/pages/index.astro",{"propagation":"none","containsHead":true}],["/home/dan/all-for-water/src/pages/marca/[marca].astro",{"propagation":"none","containsHead":true}]],"renderers":[],"clientDirectives":[["idle","(()=>{var l=(n,t)=>{let i=async()=>{await(await n())()},e=typeof t.value==\"object\"?t.value:void 0,s={timeout:e==null?void 0:e.timeout};\"requestIdleCallback\"in window?window.requestIdleCallback(i,s):setTimeout(i,s.timeout||200)};(self.Astro||(self.Astro={})).idle=l;window.dispatchEvent(new Event(\"astro:idle\"));})();"],["load","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).load=e;window.dispatchEvent(new Event(\"astro:load\"));})();"],["media","(()=>{var n=(a,t)=>{let i=async()=>{await(await a())()};if(t.value){let e=matchMedia(t.value);e.matches?i():e.addEventListener(\"change\",i,{once:!0})}};(self.Astro||(self.Astro={})).media=n;window.dispatchEvent(new Event(\"astro:media\"));})();"],["only","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).only=e;window.dispatchEvent(new Event(\"astro:only\"));})();"],["visible","(()=>{var a=(s,i,o)=>{let r=async()=>{await(await s())()},t=typeof i.value==\"object\"?i.value:void 0,c={rootMargin:t==null?void 0:t.rootMargin},n=new IntersectionObserver(e=>{for(let l of e)if(l.isIntersecting){n.disconnect(),r();break}},c);for(let e of o.children)n.observe(e)};(self.Astro||(self.Astro={})).visible=a;window.dispatchEvent(new Event(\"astro:visible\"));})();"]],"entryModules":{"\u0000@astrojs-ssr-adapter":"_@astrojs-ssr-adapter.mjs","\u0000noop-middleware":"_noop-middleware.mjs","\u0000@astro-page:src/pages/es/marca/[marca]@_@astro":"pages/es/marca/_marca_.astro.mjs","\u0000@astro-page:src/pages/marca/[marca]@_@astro":"pages/marca/_marca_.astro.mjs","\u0000@astrojs-ssr-virtual-entry":"entry.mjs","\u0000@astro-page:src/pages/contact@_@astro":"pages/contact.astro.mjs","\u0000@astro-page:src/pages/es/contacto@_@astro":"pages/es/contacto.astro.mjs","\u0000@astro-page:src/pages/[id]@_@astro":"pages/_id_.astro.mjs","\u0000@astro-renderers":"renderers.mjs","\u0000@astro-page:src/pages/es/[id_es]@_@astro":"pages/es/_id_es_.astro.mjs","\u0000@astro-page:src/pages/es/index@_@astro":"pages/es.astro.mjs","\u0000@astro-page:src/pages/index@_@astro":"pages/index.astro.mjs","\u0000@astro-page:node_modules/astro/dist/assets/endpoint/generic@_@js":"pages/_image.astro.mjs","/home/dan/all-for-water/node_modules/astro/dist/assets/services/sharp.js":"chunks/sharp_C2KTFD_Z.mjs","\u0000@astrojs-manifest":"manifest_DNVqlSOt.mjs","@components/ui/SearchComponent_es":"_astro/SearchComponent_es.CiZswkzw.js","/home/dan/all-for-water/src/components/ui/SearchComponentAside_es":"_astro/SearchComponentAside_es.BVv8fVg3.js","@components/ui/SearchComponent":"_astro/SearchComponent.CSvGPFE1.js","/home/dan/all-for-water/src/components/ui/SearchComponentAside":"_astro/SearchComponentAside.DXLSfFbe.js","@components/ui/LenguageSwitch":"_astro/LenguageSwitch.DEZ0SGRk.js","/home/dan/all-for-water/src/components/ui/ProductsDropdown_es":"_astro/ProductsDropdown_es.B39rrq2l.js","/home/dan/all-for-water/src/components/ui/AsideProductsDropdown_es":"_astro/AsideProductsDropdown_es.D5sKGcTc.js","/home/dan/all-for-water/src/components/ui/ProductsDropdown":"_astro/ProductsDropdown.BNglfWmw.js","/home/dan/all-for-water/src/components/ui/AsideProductsDropdown":"_astro/AsideProductsDropdown.CRwEP1xS.js","@astrojs/react/client.js":"_astro/client.BSOY9D2A.js","/home/dan/all-for-water/src/layouts/Layout.astro?astro&type=script&index=0&lang.ts":"_astro/Layout.astro_astro_type_script_index_0_lang.B3VmyUwG.js","/home/dan/all-for-water/src/layouts/Layout_es.astro?astro&type=script&index=0&lang.ts":"_astro/Layout_es.astro_astro_type_script_index_0_lang.B3VmyUwG.js","/home/dan/all-for-water/node_modules/astro/components/ClientRouter.astro?astro&type=script&index=0&lang.ts":"_astro/ClientRouter.astro_astro_type_script_index_0_lang.BgCNIpch.js","astro:scripts/page.js":"_astro/page.BNYwb576.js","astro:scripts/before-hydration.js":""},"inlinedScripts":[["/home/dan/all-for-water/src/layouts/Layout.astro?astro&type=script&index=0&lang.ts","document.addEventListener(\"contextmenu\",e=>{e.target.tagName===\"IMG\"&&(e.preventDefault(),alert(\"No puedes guardar esta imagen.\"))});"],["/home/dan/all-for-water/src/layouts/Layout_es.astro?astro&type=script&index=0&lang.ts","document.addEventListener(\"contextmenu\",e=>{e.target.tagName===\"IMG\"&&(e.preventDefault(),alert(\"No puedes guardar esta imagen.\"))});"]],"assets":["/_astro/_id_.Dhwt1lP-.css","/favicon.png","/_astro/AsideProductsDropdown.CRwEP1xS.js","/_astro/AsideProductsDropdown_es.D5sKGcTc.js","/_astro/ClientRouter.astro_astro_type_script_index_0_lang.BgCNIpch.js","/_astro/LenguageSwitch.DEZ0SGRk.js","/_astro/ProductsDropdown.BNglfWmw.js","/_astro/ProductsDropdown_es.B39rrq2l.js","/_astro/SearchComponent.CSvGPFE1.js","/_astro/SearchComponentAside.DXLSfFbe.js","/_astro/SearchComponentAside_es.BVv8fVg3.js","/_astro/SearchComponent_es.CiZswkzw.js","/_astro/categories.CN_oQO_X.js","/_astro/categories_es.Cx8o0Ohf.js","/_astro/client.BSOY9D2A.js","/_astro/index.Bel2lkwj.js","/_astro/index.DHQ29l5n.js","/_astro/index.Dtfyuy1a.js","/_astro/jsx-runtime.BqEXxvvU.js","/_astro/page.BNYwb576.js","/_astro/pocketbase.es.awTBJJAN.js","/fotos/agua.jpg","/fotos/espana.png","/fotos/family_water.jpg","/fotos/logo.png","/fotos/oceano.jpg","/fotos/usa.png","/_astro/page.BNYwb576.js"],"buildFormat":"directory","checkOrigin":true,"serverIslandNameMap":[],"key":"t65eYQehZYHbXWFMa+jl9XqP/xlLLv8yBIIx5HHbehI="});
+if (manifest.sessionConfig) manifest.sessionConfig.driverModule = null;
+
+export { manifest };
