+++ .vercel/output/config.json
@@ -24,7 +24,7 @@
 			"dest": "_render"
 		},
 		{
-			"src": "^\\/es\\/([^/]+?)\\/?$",
+			"src": "^\\/es\\/marca\\/([^/]+?)\\/?$",
 			"dest": "_render"
 		},
 		{
