+++ astro.config.mjs
@@ -43,7 +43,7 @@ export default defineConfig({
   //     enabled: true,
   //   }
   // }),
-  output: "hybrid",
+  output: "server",
 
   adapter: vercel({
     webAnalytics: {
