+++ src/pages/es/marca/[marca].astro
